Add files via upload
The system allows the fine tuning of an optical fiber position on an input grating, controlling its XYZ position and its angle of incidence. A similar configuration is used for the output extraction, while a fiber array can be used an alternative to monitor multiple output channels. Such a configuration can be useful for characterization of arrayed sensing devices or Optical Phased Arrays. The system can be used in the visible range and in the IR spectrum as well, provided the input light source is changed. All the setting are controlled manually, and it is prepared to easily perform a sequential analysis of a set of identical circuits, fabricated in a parallel configuration on the same chip.